Use the payload param if Github sent it
When looking at the logs, two things can be sent in a Github hook payload. Either the payload is sent in a way that rails parses it and puts it in the params, or there is a payload parameter with stringified json as its value. (Presumably this depends on which content-type people selected the hook to be sent as.)