Fix exception notifications in delayed job failure handler
The first argument of `notify_exception` has to be an exception, which `job.last_error` isn't. A dummy exception is passed, since I couldn't find a way to create an exception from `job.last_error`.

This also makes the notifier run in development, since there is not really a reason to only make it run in production. If an exception happens in development it's nice to know about it (and it also helps with debugging).