Change how submissions are filtered by course and exercise on the exercise show page
Previously, we assumed that the user was on the right page and just requested a relative URL. When a user submitted from a an exercise page within a course and was not a member of the course this led to problems. These problems were fixed in 5049cedec69101bc888632fd9f3a2a5e5efbad26. However, this fix introduced #884. This change fixes #884 without reintroducing the problems of #718. It does so by requesting `submissions.js` from the root and adding the right parameters.