Fix pagination of submissions on user-specific submissions index
Because of the changes in ec83e014 (#1103) the user id is explicitly set to nil, causing the `/user/<id>/` part of the URL to be stripped on the user-specific submissions index. This is fixed by only setting `user_id` in the params if we have a user to get the id from.