Fix users being unable to subscribe to hidden courses
The condition for adding the secret token to the URL was wrong, resulting in the secret not being added on courses that were open and hidden. This fixes the condition for the secret. The condition for the checking of the secret was also wrong, so I updated that one as well.