Fix synchronization issues around timeout detection
It seems that in most (all?) cases, the main thread continues when `container.stop` is called from the timer thread. This results in the timer thread not having finished yet (and therefore not having set its value to `true`). This made us misreport the reason the container was stopped.

By synchronizing around a mutex we can be sure the timer thread gets to set the `timeout` value if it was the reason the container stopped.