Add `stub_all_exercises!` stub helper
We create a lot of exercises which are not backed by files in our test
cases. Instead we stub some of their methods so they act like there are
underlying files (config.json and description).

However, when an exercise is recovered from the database, these stubs
are not active anymore because it is a new Exercise object. Resulting in
Exercises which are not valid anymore (because their files don't exist).

This commit adds a helper which stubs all exercise objects.