Allow multiple code listings
The code listing javascript is only used when there are annotations.
To be able to copy from the exercise info page, I've split of the
copy functionality: it's now loaded regardless of annotations. There
is now also support for multiple instances of the feedback code on the
same page (regarding copying, annotations still won't work probably).