Series factory: create content_pages and exercises
Change the behavior of the activities_count parameter to create both
exercises and series (in case of an odd count, the series will have one
exercise more than it has content_pages).

Two parameters content_page_count and exercises_count were also added to
explicitly specify how much of each activity type should be created.

The course factory was also changed with the same behavior, which is
delegated to the series.

The tests are updated with this behavior.