Fix double border in code blocks in global annotations
In the code listing itself, the `pre` elements have some styling applied
to them. This is also applied to the `pre` inside annotations (probably
unintended). Resolved by duplicating the relevant styles for `pre`
elements inside annotations.

This does duplicate the styles, but as it concerns two semantically
different elements, I don't think that's an issue.

Fixes #2224.