Prefer break-spaces when available
Very long uninterrupted sequences of spaces cause the line to overflow
in Blink-based browsers (Chrome, Edge). This is not an issue on Firefox.
Not sure if a browser bug or intended behaviour (the spec is vague).

Using the newer attribute "break-spaces" fixes this in Chrome and has
no effect in Firefox, so it seems like a good solution to me.

More information:
https://www.chromestatus.com/feature/4842781306519552
https://developer.mozilla.org/en-US/docs/Web/CSS/white-space

Note: I've left the previous value as fallback, since not all browsers
seem to support it. In particular, Mozilla claims Safari doesn't support
it, yet Chrome claims it does.

Fixes #2221