Always init masonry roots
The load event isn't triggered when returning from another page or when the listener is installed after the event has already triggered. Both things happened regularly on my machine (Firefox on Linux). This pull request moves the initialisation of the roots from the event handler to the init function of the homepage javascript. This fixes all issues for me. I also tested chromium on my machine, everything still works as expected there as well.