Fix feedback table backgrounds in dark mode
The dark theme for Rouge also specified a background for the code. In
most places, this was already overridden by other styles, except for the
feedback table. As a solution I've just removed the first background
declaration; the light theme also doesn't have this, and it doesn't seem
to break anything.

Fixes #2337.