Transform exercise on submission to activity
Since exercises can be converted to content pages, we can no longer
assume every submission has an exercise; every submission does have an
activity.

Wherever possible, the use of exercise was changed to use the activity
instead. When not possible (e.g. judging), the code was made nil-proof
or fails with an error message.

See the linked issue for more information and considerations.

Linked issue: #2364