Add retry mechanism to docker start
Starting the docker container sometimes fails with the message "docker
read timeout reached". This seems to be a temporary failure due to
high load. This adds a retry mechanism where the docker container is
started again after a second. If it still fails at that point, we go
back to the previous behaviour.